+++ lma-ai-stack/source/ui/src/components/meetings-query-layout/breadcrumbs.jsx
@@ -5,7 +5,7 @@
  */
 import React from 'react';
 
-import { BreadcrumbGroup } from '@awsui/components-react';
+import { BreadcrumbGroup } from '@cloudscape-design/components';
 
 import { MEETINGS_QUERY_PATH, DEFAULT_PATH } from '../../routes/constants';