+++ lma-ai-stack/source/ui/src/components/transcript-summary-config/breadcrumbs.jsx
@@ -4,7 +4,7 @@
  * See the LICENSE file in the project root for full license information.
  */
 import React from 'react';
-import { BreadcrumbGroup } from '@awsui/components-react';
+import { BreadcrumbGroup } from '@cloudscape-design/components';
 import { TRANSCRIPT_SUMMARY_PATH, DEFAULT_PATH } from '../../routes/constants';
 
 export const transcriptSummaryBreadcrumbItems = [